requirements.txt
PyMuPDF
frontend
fitz
Pillow